# Add schema pruning performance benchmarks