--- a/bench/src/test/scala/org/apache/spark/sql/protobuf/backport/jmh/ProtobufConversionJmhBenchmarkDom.scala
+++ b/bench/src/test/scala/org/apache/spark/sql/protobuf/backport/jmh/ProtobufConversionJmhBenchmarkDom.scala
@@ -201,7 +201,7 @@ class ProtobufConversionJmhBenchmarkDom {
     bh.consume(domPrunedParser.parse(deepDomBinary))
   }
 
-  // @Benchmark
+  @Benchmark
   def domDeepProtoParsing(bh: Blackhole): Unit = {
     bh.consume(DomDocument.parseFrom(deepDomBinary))
   }