## Summary

Adds JMH benchmarks to measure schema pruning performance for flat, nested, and complex recursive schemas.

**Flat schema benchmark** (`simplePrunedWireFormatParser`):
- Accesses single scalar field (field_double_055) from 120-field message
- **4.6x speedup**: 1,729ns → 379ns (78% reduction)

**Nested schema benchmark** (`domDeepPrunedWireFormatParser`):
- Accesses deeply nested field (root.children[].children[].children[].tag_name) from recursive DOM structure
- **12.6x speedup**: 30,082ns → 2,380ns (92% reduction)

**Complex schema benchmark** (`complexPrunedWireFormatParser`):
- Accesses nested scalar field (message_b.nested_data.count) from complex recursive structure
- **8.0x speedup**: 1,941ns → 241ns (88% reduction)

## Key Findings

- Schema pruning delivers 4.5x+ speedup for single-field access on wide schemas
- Nested field access shows compound benefits (8-12x) by skipping entire subtrees at each level
- Performance gains scale with schema width and nesting depth
- Deeper nesting (DOM: 4 levels) shows larger speedup than moderate nesting (Complex: 3 levels)

## Test Plan

Run benchmarks:
```bash
sbt 'bench/Jmh/run .*ProtobufConversionJmhBenchmarkSimple.*'
sbt 'bench/Jmh/run .*ProtobufConversionJmhBenchmarkDom.*'
sbt 'bench/Jmh/run .*ProtobufConversionJmhBenchmarkComplex.*'
```

🤖 Generated with [Claude Code](https://claude.com/claude-code)